+++ src/planning/PlannerConfigurationSelector.ts
@@ -4,7 +4,7 @@
  * ------------------------------------------------------------------------------------------ */
 import { window, Uri, QuickPickItem, workspace } from 'vscode';
 import * as path from 'path';
-import { AsyncServiceConfiguration, PlannerAsyncService } from 'pddl-planning-service-client';
+import { AsyncServiceOnlyConfiguration, PlannerAsyncService } from 'pddl-planning-service-client';
 
 export class PlannerConfigurationSelector {
 
@@ -45,7 +45,7 @@ export class PlannerConfigurationSelector {
         return selectedUris[0];
     }
 
-    static async loadConfiguration(configurationUri: Uri, timeout: number): Promise<AsyncServiceConfiguration> {
+    static async loadConfiguration(configurationUri: Uri, timeout: number): Promise<AsyncServiceOnlyConfiguration> {
         if (configurationUri.toString() === PlannerConfigurationSelector.DEFAULT.toString()) {
             return PlannerAsyncService.createDefaultConfiguration(timeout);
         }